funds-manager: Reorganize api types and helpers ahead of execution impl
