+++ funds-manager/funds-manager-api/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2021"
 [dependencies]
 
 renegade-api = { package = "external-api", workspace = true }
+
+ethers = "2"
+hex = "0.4.3"
 hmac = "0.12.1"
 http = "0.2.12"
 itertools = "0.13.0"
@@ -13,3 +16,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sha2 = "0.10.7"
 uuid = "1.7.1"
+
+[dev-dependencies]
+rand = "0.8.5"
