+++ funds-manager/funds-manager-api/src/types/mod.rs
@@ -1,0 +1,9 @@
+//! API types for the funds manager
+
+pub mod fees;
+pub mod gas;
+pub mod hot_wallets;
+pub mod quoters;
+
+/// The ping route
+pub const PING_ROUTE: &str = "ping";
